Move to frame based timing for display, smaller datatype for cycle updates

+++ elbgb.gameboy/SerialIO.cs
@@ -14,7 +14,7 @@ namespace elbgb.gbcore
 
 		}
 
-		public override void Update(ulong cycleCount)
+		public override void Update(uint cycleCount)
 		{
 			return;
 		}
